server/requirements.txt
# FastAPI and Core
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Google Cloud Platform and Vertex AI
google-cloud-aiplatform
google-generativeai
google-cloud-storage
google-auth
google-auth-oauthlib
google-auth-httplib2
google.adk

# Database
pg8000
asyncpg
sqlalchemy
alembic

# Vector Search and Embeddings
sentence-transformers
numpy
scipy

# Document Processing
PyPDF2
python-docx
python-multipart
aiofiles

# LLM Integration
openai

# HTTP Client
httpx
requests

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]

# Caching & Background Tasks
redis

# ALM Integration
jira
azure-devops
PyGithub
atlassian-python-api

# MCP Tools (optional)
mcp-atlassian

# Utilities
python-dotenv
structlog
rich

# Testing
pytest
pytest-asyncio

# Development
black
isort
flake8
mypy